Replacing a static class with fluent methods
